Migrated doc generator to PyccoPlus and included XResources
This included a complete rewrite of the Makefile, which is now more
streamlined. In addition, the CSS for the code view is no longer Docco's
but based on Pycco's, which is tableless. The animation for
collapse/uncollapse therefore also includes the fixed background where
the code sits. Due to switching to PyccoPlus, dircolors and conky files
have been removed as they will require additional lexers and the
makeshift bash lexer used before cannot be specified on-the-fly.